requirements.txt
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-cors-headers==3.3.0
django-easy-pdf==0.1.1
django-filter==2.2.0
django-import-export==2.1.0
django-rest-passwordreset==1.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
geographiclib==1.50
geopy==1.22.0
html5lib==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
PyJWT==1.7.1
PyPDF2==1.26.0
python-barcode==0.11.0
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.42
six==1.14.0
sqlparse==0.3.1
tablib==2.0.0
url==0.4.2
webencodings==0.5.1
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0